modified build script

+++ build.sh
@@ -1,14 +1,15 @@
 #!/bin/bash
 set -e
+type bash
 
 SERVICE_REPO_DIR=git_repos
-DOCKER_REGISTRY_HOST=panosc-docker.ill.fr
 
+read -p "What is the address for your registry? " -r DOCKER_REGISTRY_HOST
 
 # Create and clean git repos dir
 mkdir -p $SERVICE_REPO_DIR
-rm -rf $SERVICE_REPO_DIR/*
-rm -rf $SERVICE_REPO_DIR/.* ||:
+rm -rf  "${SERVICE_REPO_DIR:?}/"*
+rm -rf "${SERVICE_REPO_DIR:?}/".* ||:
 
 
 git clone https://github.com/panosc-portal/cloud-provider-kubernetes.git $SERVICE_REPO_DIR/cloud-provider-kubernetes
@@ -24,4 +25,4 @@ docker push $DOCKER_REGISTRY_HOST/panosc-portal/cloud-provider-kubernetes
 ## Remove git repos dir
 rm -rf $SERVICE_REPO_DIR
 
-helm install panosc-portal-demo ./panosc-portal-demo
+helm install panosc-portal-demo  --set dockerRegistry.host="${DOCKER_REGISTRY_HOST}"   ./panosc-portal-demo
